requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.2.0
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
